agstack token error typo fix; tidy_tables.py renamed risk.py; removed unused modules such as exporting to a ceo project
